libs/racing-coach-api-client/pyproject.toml
[project]
name = "racing-coach-server-client"
version = "0.1.0"
description = "A client library for accessing Racing Coach Server"
authors = [{ name = "Sawyer", email = "sawyerczupka@gmail.com" }]
requires-python = ">=3.12"
readme = "README.md"
dependencies = [
    "httpx>=0.23.0,<0.29.0",
    "attrs>=22.2.0",
    "python-dateutil>=2.8.0,<3",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["racing_coach_server_client"]

[dependency-groups]
dev = [
    "openapi-python-client>=0.27.1",
]

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = ["F", "I", "UP"]
